precast _max_hash and _mersenne_prime to uint64

+++ datasketch/minhash.py
@@ -8,8 +8,8 @@ from datasketch.hashfunc import sha1_hash32
 hashvalue_byte_size = len(bytes(np.int64(42).data))
 
 # http://en.wikipedia.org/wiki/Mersenne_prime
-_mersenne_prime = (1 << 61) - 1
-_max_hash = (1 << 32) - 1
+_mersenne_prime = np.uint64((1 << 61) - 1)
+_max_hash = np.uint64((1 << 32) - 1)
 _hash_range = (1 << 32)
 
 class MinHash(object):
@@ -131,7 +131,7 @@ class MinHash(object):
         '''
         hv = self.hashfunc(b)
         a, b = self.permutations
-        phv = np.bitwise_and((a * hv + b) % _mersenne_prime, np.uint64(_max_hash))
+        phv = np.bitwise_and((a * hv + b) % _mersenne_prime, _max_hash)
         self.hashvalues = np.minimum(phv, self.hashvalues)
 
     def jaccard(self, other):
